feat: add decimal.

+++ src/ChampionsChromo.Application/Albums/Commands/UpdateAlbum/UpdateAlbumCommand.cs
@@ -14,7 +14,7 @@ public record UpdateAlbumCommmand : IRequest<Result>
     public bool HasLegend { get; set; }
     public bool HasA4 { get; set; }
     public int TotalStickers { get; set; }
-    public int CommonPrice { get; set; }
-    public int LegendPrice { get; set; }
-    public int A4Price { get; set; }
+    public decimal CommonPrice { get; set; }
+    public decimal LegendPrice { get; set; }
+    public decimal A4Price { get; set; }
 }
