+++ src/ChampionsChromo.Core/Entities/School.cs
@@ -11,5 +11,5 @@ public class School : Entity
     public string State { get; init; } = string.Empty;
     public string Warning { get; set; } = string.Empty;
     public string BgWarningColor { get; set; } = string.Empty;
-    public int ShippingCost { get; set; }
+    public decimal ShippingCost { get; set; }
 }
