+++ src/ChampionsChromo.Core/Models/UpdateSchoolDto.cs
@@ -8,7 +8,7 @@
         public string Warning { get; set; } = string.Empty;
         public string BgWarningColor { get; set; } = string.Empty;
         public string ImageUrl { get; set; } = string.Empty;
-        public int ShippingCost { get; set; }
+        public decimal ShippingCost { get; set; }
 
     }
 }
